update error state
